wip: refactor api (modules groups and translations) and add new constants for centralize commons values
