+++ api/modules/constants.js
@@ -3,5 +3,9 @@ let path = require('path')
 module.exports = Object.freeze({
   JSON_NB_SPACES_INDENT: 2,
   PATH_API: '/api',
-  PATH_JSON_FOLDER: path.join(__dirname, '../json')
+  PATH_JSON_FOLDER: path.join(__dirname, '../json'),
+  DEFAULT_PATH_JSON_FOLDER: path.join(__dirname, '../json'),
+  ADD: 'add',
+  UPDATE: 'update',
+  DELETE: 'delete'
 })
